src/common/queue/queue-linked-list.test.ts
import { Queue } from './queue-linked-list';

describe('Queue', () => {
  const queue = new Queue();

  beforeEach(() => queue.empty());

  it('should create empty queue', () => {
    expect(queue).not.toBeNull();
  });

  it('enqueue()', () => {
    queue.enqueue(1);
    queue.enqueue(2);

    expect(queue.toString()).toBe('1,2');
  });

  it('peek()', () => {
    expect(queue.peek()).toBeNull();

    queue.enqueue(1);
    queue.enqueue(2);

    expect(queue.peek()).toBe(1);
    expect(queue.peek()).toBe(1);
  });

  it('isEmpty()', () => {
    expect(queue.isEmpty()).toBe(true);

    queue.enqueue(1);

    expect(queue.isEmpty()).toBe(false);
  });

  it('should dequeue from queue', () => {
    queue.enqueue(1);
    queue.enqueue(2);

    expect(queue.dequeue()).toBe(1);
    expect(queue.dequeue()).toBe(2);
    expect(queue.dequeue()).toBeNull();
    expect(queue.isEmpty()).toBe(true);
  });
});
